+++ test/FormActionHandlers.test.js
@@ -8,14 +8,14 @@ var TestUtils$ResX = require("./TestUtils.js");
 Buntest.describe("Form action handlers", (function () {
         Buntest.test("prefixing of form action handler routes work", (async function () {
                 Handlers$ResX.formAction(TestUtils$ResX.Handler.handler, "/test", (async function (param) {
-                        return new Response("Test!", undefined);
+                        return new Response("Test!");
                       }));
                 var response = await TestUtils$ResX.getResponse((function (param) {
                         return "nope";
                       }), undefined, "/_form/test");
                 var text = await response.text();
                 Buntest.expect(text).toBe("Test!");
-              }), undefined);
+              }));
       }));
 
 /*  Not a pure module */