+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # master
 
+- Add `allowEmptyString` option to `FormDataHelpers.getString`, and make sure that empty strings are treated as `None` by default.
+- Handle `on` and `off` values for `FormDataHelpers.getBool`.
+
 # 0.1.1
 
 - Fix issue with Vite transform and CSS files with imports.